requirements_flask.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy>=1.24.0
xgboost>=2.0.0
prophet>=1.1.5
scikit-learn>=1.3.0
